# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API文档",
        "title": "API文档接口",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/user/del": {
            "post": {
                "description": "删除用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关接口"
                ],
                "summary": "删除用户接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "查询参数",
                        "name": "params",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controllers.delUser"
                        }
                    }
                ]
            }
        },
        "/user/getlist": {
            "post": {
                "description": "删除用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关接口"
                ],
                "summary": "获取用户接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "params",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/controllers.queryUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/save": {
            "post": {
                "description": "添加用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关接口"
                ],
                "summary": "添加用户接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "查询参数",
                        "name": "params",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ]
            }
        }
    },
    "definitions": {
        "controllers.delUser": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "controllers.queryUser": {
            "type": "object",
            "properties": {
                "page": {
                    "description": "页码",
                    "type": "integer"
                },
                "pagesize": {
                    "description": "每页数量",
                    "type": "integer"
                },
                "role_name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "required": [
                "password",
                "role_id",
                "role_name",
                "username"
            ],
            "properties": {
                "created_at": {
                    "description": "创建时间，不用传",
                    "type": "string"
                },
                "id": {
                    "description": "id",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "role_id": {
                    "description": "角色id",
                    "type": "string"
                },
                "role_name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间，不用传",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        }
    }
}